+++ designer/src/toolbox/toolbox-extension.ts
@@ -4,6 +4,6 @@ import { Toolbox } from './toolbox';
 
 export class ToolboxExtension implements UiComponentExtension {
 	public create(root: HTMLElement, api: DesignerApi): UiComponent {
-		return Toolbox.create(root, api.toolbox);
+		return Toolbox.create(root, api.toolbox, api.i18n);
 	}
 }